organize cron code and fix collabland
