+++ apps/cron/src/tasks/processGithubWebhookMessages/index.ts
@@ -1,5 +1,6 @@
 import { log } from '@charmverse/core/log';
-import { processMessages } from '@root/lib/aws/webhookSqs';
+
+import { processMessages } from '../../webhookSqs';
 
 import { SQS_WEBHOOK_GITHUB_QUEUE_NAME } from './constants';
 import { processWebhookMessage } from './webhook/processWebhookMessage';
@@ -9,5 +10,5 @@ const queueUrl = SQS_WEBHOOK_GITHUB_QUEUE_NAME || '';
 
 export async function task() {
   log.debug('Starting to process github webhook messages job');
-  processMessages({ processorFn: processWebhookMessage, queueUrl });
+  await processMessages({ processorFn: processWebhookMessage, queueUrl });
 }
