+++ apps/cron/src/tasks/processMailgunWebhookMessages/task.ts
@@ -1,7 +1,8 @@
 import { log } from '@charmverse/core/log';
-import { processMessages } from '@root/lib/aws/webhookSqs';
 import { SQS_WEBHOOK_MAILGUN_QUEUE_NAME } from '@root/lib/mailer/mailgunClient';
 import { processWebhookMessage } from '@root/lib/mailer/webhook/processWebhookMessage';
+
+import { processMessages } from '../../webhookSqs';
 
 log.info('Mailgun Queue url:', SQS_WEBHOOK_MAILGUN_QUEUE_NAME);
 const queueUrl = SQS_WEBHOOK_MAILGUN_QUEUE_NAME || '';
@@ -12,5 +13,5 @@ export async function task() {
     return;
   }
   log.debug('Starting to process mailgun webhook messages job');
-  processMessages({ processorFn: processWebhookMessage, queueUrl });
+  await processMessages({ processorFn: processWebhookMessage, queueUrl });
 }
